Move AzureClientError and update copyright strings
